implement authentication

+++ controllers/candidate.js
@@ -1,6 +1,7 @@
 const Candidate = require("../models/candidate");
 const CourtSearch = require("../models/court-search");
 const AdverseAction = require("../models/adverse-action");
+const User = require("../models/user");
 
 exports.getCandidates = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -17,7 +18,34 @@ exports.getCandidates = (req, res, next) => {
     .then((candidates) => {
       return res.status(200).json({
         message: "Fetched candidates successfully.",
-        candidates: candidates,
+        data: candidates,
+        totalItems: totalItems,
+      });
+    })
+    .catch((err) => {
+      if (!err.statusCode) {
+        err.statusCode = 500;
+      }
+      next(err);
+    });
+};
+
+exports.getCandidatesByUser = (req, res, next) => {
+  const currentPage = req.query.page || 1;
+  const perPage = req.query.perPage || 2;
+  let totalItems;
+  Candidate.find({ user: req.userId })
+    .countDocuments()
+    .then((count) => {
+      totalItems = count;
+      return Candidate.find({ user: req.userId })
+        .skip((currentPage - 1) * perPage)
+        .limit(perPage);
+    })
+    .then((candidates) => {
+      return res.status(200).json({
+        message: "Fetched candidates successfully.",
+        data: candidates,
         totalItems: totalItems,
       });
     })
@@ -38,9 +66,14 @@ exports.getCandidate = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (candidate.user.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
       return res
         .status(200)
-        .json({ message: "candidate fetched.", candidate: candidate });
+        .json({ message: "candidate fetched.", data: candidate });
     })
     .catch((err) => {
       if (!err.statusCode) {
@@ -64,14 +97,22 @@ exports.createCandidate = (req, res, next) => {
     dob: dob,
     location: location,
     reports: reports,
+    user: req.userId,
   });
   candidate
     .save()
     .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.candidates.push(candidate);
+      return user.save();
+    })
+    .then((result) => {
       console.log("Created Candidate");
       return res
         .status(201)
-        .json({ message: "Candidate created successfully", candidate: result });
+        .json({ message: "Candidate created successfully" });
     })
     .catch((err) => {
       if (!err.statusCode) {
@@ -90,8 +131,22 @@ exports.deleteCandidate = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (candidate.user.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
       await CourtSearch.deleteMany({ candidateId: candidateId });
       await AdverseAction.deleteMany({ candidate: candidateId });
+    })
+    .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.candidates.pull(candidateId);
+      return user.save();
+    })
+    .then((result) => {
       return res.status(200).json({ message: "candidate deleted." });
     })
     .catch((err) => {
@@ -111,8 +166,15 @@ exports.getReportByCandId = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
-      const r = candidate.reports;
-      return res.status(200).json({ message: "report fetched.", reports: r });
+      if (candidate.user.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
+      const candidateReports = candidate.reports;
+      return res
+        .status(200)
+        .json({ message: "report fetched.", data: candidateReports });
     })
     .catch((err) => {
       if (!err.statusCode) {
@@ -129,6 +191,11 @@ exports.updateReportByCandId = (req, res, next) => {
       if (!candidate) {
         const error = new Error("Could not find candidate.");
         error.statusCode = 404;
+        throw error;
+      }
+      if (candidate.user.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
         throw error;
       }
       const report = candidate.reports;
